@@ -1,49 +1,73 @@
 package com.example;
 
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+//import javax.servlet.ServletException;
+//import javax.servlet.http.HttpServlet;
+//import javax.servlet.http.HttpServletRequest;
+//import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse; 
 import java.io.IOException;
 import java.sql.Connection;
-import java.sql.DriverManager;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
+import java.sql.SQLException;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
-@WebServlet("/login")
 public class LoginServlet extends HttpServlet {
     private static final long serialVersionUID = 1L;
+    private static final Logger logger = LogManager.getLogger(LoginServlet.class);
 
+    @Override
+    public void init() throws ServletException {
+        try (Connection conn = DatabaseUtil.getConnection()) {
+            logger.info("Database connection successful");
+        } catch (SQLException e) {
+            logger.error("Database connection failed: " + e.getMessage(), e);
+            throw new ServletException("Database connection failed", e);
+        }
+    }
+
+    @Override
+    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        response.setContentType("text/html");
+        response.getWriter().println("<h1>Login Page</h1>");
+        response.getWriter().println("<form method='post'>");
+        response.getWriter().println("Username: <input type='text' name='username'><br>");
+        response.getWriter().println("Password: <input type='password' name='password'><br>");
+        response.getWriter().println("<input type='submit' value='Login'>");
+        response.getWriter().println("</form>");
+    }
+
+    @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         String username = request.getParameter("username");
         String password = request.getParameter("password");
 
+        response.setContentType("text/html");
         if (authenticate(username, password)) {
+            logger.info("Successful login attempt for user: {}", username);
             response.getWriter().println("<h2>Login Successful!</h2>");
         } else {
+            logger.warn("Failed login attempt for user: {}", username);
             response.getWriter().println("<h2>Login Failed!</h2>");
         }
     }
 
     private boolean authenticate(String username, String password) {
-        boolean isAuthenticated = false;
-        String jdbcUrl = "jdbc:postgres://localhost:5432/tododb";
-        String dbUser = "user1";
-        String dbPassword = "changeme";
-
-        try (Connection conn = DriverManager.getConnection(jdbcUrl, dbUser, dbPassword)) {
-            String sql = "SELECT * FROM users WHERE username = ? AND password = ?";
-            try (PreparedStatement stmt = conn.prepareStatement(sql)) {
-                stmt.setString(1, username);
-                stmt.setString(2, password);
-                try (ResultSet rs = stmt.executeQuery()) {
-                    isAuthenticated = rs.next();
-                }
+        String sql = "SELECT * FROM users WHERE username = ? AND password = ?";
+        try (Connection conn = DatabaseUtil.getConnection();
+             PreparedStatement stmt = conn.prepareStatement(sql)) {
+            stmt.setString(1, username);
+            stmt.setString(2, password);
+            try (ResultSet rs = stmt.executeQuery()) {
+                return rs.next();
             }
-        } catch (Exception e) {
-            e.printStackTrace();
+        } catch (SQLException e) {
+            logger.error("Authentication error: " + e.getMessage(), e);
+            return false;
         }
-        return isAuthenticated;
     }
 }
