@@ -5,12 +5,31 @@ import java.sql.DriverManager;
 import java.sql.SQLException;
 
 public class DatabaseUtil {
-    private static final String JDBC_URL = "jdbc:postgres://localhost:5432/tododb";
+    private static final String JDBC_URL = "jdbc:postgresql://chetushree-db1-1:5432/tododb";
     private static final String DB_USER = "user1";
     private static final String DB_PASSWORD = "changeme";
 
+    static {
+        try {
+            Class.forName("org.postgresql.Driver");
+            System.out.println("PostgreSQL JDBC Driver loaded successfully");
+        } catch (ClassNotFoundException e) {
+            System.err.println("PostgreSQL JDBC Driver not found: " + e.getMessage());
+            throw new ExceptionInInitializerError(e);
+        }
+    }
+
     public static Connection getConnection() throws SQLException {
-        return DriverManager.getConnection(JDBC_URL, DB_USER, DB_PASSWORD);
+        System.out.println("Attempting to connect to database...");
+        System.out.println("JDBC URL: " + JDBC_URL);
+        System.out.println("User: " + DB_USER);
+        try {
+            Connection conn = DriverManager.getConnection(JDBC_URL, DB_USER, DB_PASSWORD);
+            System.out.println("Database connection established successfully");
+            return conn;
+        } catch (SQLException e) {
+            System.err.println("Failed to connect to database. Error: " + e.getMessage());
+            throw e;
+        }
     }
 }
-
