@@ -1,47 +1,55 @@
 package com.example;
 
 import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.remote.DesiredCapabilities;
-import org.openqa.selenium.remote.RemoteWebDriver;
+import org.openqa.selenium.WebElement;
+
+import org.openqa.selenium.support.ui.WebDriverWait;
+import org.openqa.selenium.support.ui.ExpectedConditions;
+import org.openqa.selenium.OutputType;
+import org.openqa.selenium.TakesScreenshot;
+import org.apache.commons.io.FileUtils;
+
 import org.testng.Assert;
-import org.testng.annotations.AfterClass;
-import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
+import java.io.File;
+import java.time.Duration;
+import java.util.logging.Logger;
 
-import java.net.MalformedURLException;
-import java.net.URL;
-
-public class LoginFailureTest {
-    private WebDriver driver;
-
-    @BeforeClass
-    public void setUp() throws MalformedURLException {
-        DesiredCapabilities capabilities = new DesiredCapabilities();
-        capabilities.setBrowserName("chrome");
-        
-        // Connect to the Selenium Grid hub
-        driver = new RemoteWebDriver(new URL("http://selenium-hub:4444/wd/hub"), capabilities);
-    }
+public class LoginFailureTest extends BaseTest {
+    private static final Logger logger = Logger.getLogger(LoginFailureTest.class.getName());
 
     @Test
-    public void testLoginFailure() {
-        driver.get("http://tomcat:8888/login"); // Adjust URL as needed
+    public void testFailedLogin() {
+        logger.info("Starting failed login test");
 
-        // Enter invalid username and password
-        driver.findElement(By.name("username")).sendKeys("invalidUser");
-        driver.findElement(By.name("password")).sendKeys("invalidPassword");
-        driver.findElement(By.name("submit")).click();
-        
-        // Verify that login failed
-        // Adjust the following line according to how your application shows login failures
-        String errorMessage = driver.findElement(By.id("error-message")).getText();
-        Assert.assertTrue(errorMessage.contains("Invalid credentials"), "Error message not displayed as expected.");
-    }
+        WebElement usernameField = driver.findElement(By.name("username"));
+        WebElement passwordField = driver.findElement(By.name("password"));
+        WebElement loginButton = driver.findElement(By.cssSelector("input[type='submit']"));
 
-    @AfterClass
-    public void tearDown() {
-        driver.quit();
+        usernameField.sendKeys("user1");
+        passwordField.sendKeys("invalidpassword");
+        loginButton.click();
+
+        WebDriverWait wait = new WebDriverWait(driver, Duration.ofSeconds(10));
+        WebElement failureMessage = wait.until(ExpectedConditions.presenceOfElementLocated(By.tagName("h2")));
+
+        try {
+            wait.until(ExpectedConditions.textToBe(By.tagName("h2"), "Login Failed!"));
+        } catch (Exception e) {
+            logger.warning("Expected text not found. Current text: " + failureMessage.getText());
+        }
+
+        try {
+            File screenshot = ((TakesScreenshot) driver).getScreenshotAs(OutputType.FILE);
+            FileUtils.copyFile(screenshot, new File("login_failure.png"));
+            logger.info("Screenshot saved as login_failure.png");
+        } catch (Exception e) {
+            logger.warning("Failed to save screenshot: " + e.getMessage());
+        }
+
+        logger.info("Page source after login attempt: " + driver.getPageSource());
+
+        Assert.assertEquals(failureMessage.getText(), "Login Failed!", "Login failure message is incorrect");
+        logger.info("Failed login test completed");
     }
 }
-
